Fixed various bugs

+++ Assets/Scripts/ScriptableObjects/GameplayConfig.cs
@@ -8,7 +8,7 @@ namespace ScriptableObjects
         public int minUnstability;
         public int maxUnstability = 100;
         public float unstabilityPerSecond;
-        public int workerFixScore;
+        public float workerFixScore;
         public float easyThreshold = 25;
         public float mediumThreshold = 50;
         public float hardThreshold = 75;
